Module dashu::base::ring

Trait definitions for operations related to rings (integer/polynomial/etc.)

Traits

CubicRootRem

Computer the floored cubic root of the number and return the remainder at the same time.

DivEuclid

Compute Euclidean quotient.

DivRem

Compute quotient and remainder at the same time.

DivRemAssign

Compute quotient inplace and return remainder at the same time.

DivRemEuclid

Compute Euclidean quotient and remainder at the same time.

ExtendedGcd

Compute the greatest common divisor between self and the other operand, and return both the common divisor g and the Bézout coefficients respectively.

Gcd

Compute the greatest common divisor.

RemEuclid

Compute Euclidean remainder.

SquareRootRem

Computer the floored square root of the number and return the remainder at the same time.