Struct dashu::float::FBig

pub struct FBig<RoundingMode = Zero, const BASE: u64 = 2>where
    RoundingMode: Round,{ /* private fields */ }

An arbitrary precision floating point number with arbitrary base and rounding mode.

The float number consists of a Repr and a Context. The Repr instance determines the value of the number, and the Context contains runtime information (such as precision limit, rounding mode, etc.)

For how the number is represented, see Repr, for how the precision limit and rounding mode is applied, see Context.

The arithmetic operations on FBig follows the behavior of its associated context. If a different precision limit and/or rounding mode is required, or the rounding information has to be preserved, use the methods of the Context type.

Generic Parameters

The const generic parameters will be abbreviated as BASE -> B, RoundingMode -> R. THe BASE must be in range [2, isize::MAX], and the RoundingMode can be chosen from the mode module.

With the default generic parameters, the floating number is of base 2 rounded towards zero. This is the most efficient format for operations. To represent a decimal number, the alias DBig is provided, which is base 10 rounded to the nearest.

Parsing and printing

To create a FBig instance, there are four ways:

Use predifined constants (e.g. FBig::ZERO, FBig::ONE, FBig::NEG_INFINITY).

Use the literal macro fbig! or dbig! defined in the dashu-macro crate.

Construct from the significand and exponent using from_parts() or from_parts_const().

Parse from a string.

Conversion from and to str is limited to native radix (i.e. base). To print or parse with different radix, please use to_binary(), to_decimal() or with_base(), with_base_and_precision() to convert.

For printing, currently only the Display and Debug are supported. Other formatting traits will be supported in future.

// parsing
let a = DBig::from_parts(123456789.into(), -5);
let b = DBig::from_str_native("1234.56789")?;
let c = DBig::from_str_native("1.23456789e3")?;
assert_eq!(a, b);
assert_eq!(b, c);

// printing
assert_eq!(format!("{}", DBig::from_str_native("12.34")?), "12.34");
let x = DBig::from_str_native("10.01")?
    .with_precision(0) // use unlimited precision
    .value();
if dashu_int::Word::BITS == 64 {
    // number of digits to display depends on the word size
    assert_eq!(
        format!("{:?}", x.powi(100.into())),
        "1105115697720767968..1441386704950100001 * 10 ^ -200 (prec: 0, rnd: HalfAway)"
    );
}

For detailed information of parsing, refer to the from_str_native() method.

Restriction on binary operators

Binary operators on FBig instances are restricted to the same base and same rounding mode. This is designed to make sure that no hidden conversion is performed during the operators. However, for equality test and comparsion, two FBig instances can have different rounding modes (but not different bases), because rounding will never happends during comparison.

The infinities are converted as it is, and the subnormals are converted using its actual values.

IEEE 754 behavior compliance

The representation of the floating point number doesnt follows the IEEE 754 standard, as its not designed for arbitrary precision numbers. The key differences include:

FBig doesnt support NaN values. In places where IEEE 754 operations generate NaNs, FBig will panic.

FBig doesnt have subnormal values.

FBig doesnt have negative zeros. There is only on zero value (FBig::ZERO).

Division by zero and logarithm on zero panic instead of returning infinities.

FBig operations will panic if the result overflows or underflows.

FBig does support infinities, but currently infinities are not allowed to be operated with, except for equality test and comparison.

These behaviors are subject to changes in the future.

Convert from/to f32/f64

Converting from f32/f64 to FBig is only defined for base 2 FBig (using TryFrom) to ensure the conversion is lossless. Since FBig doesnt support NANs, converting f32::NAN or f64::NAN will return Err.

Converting to f32/f64 (using to_f32() and to_f64()) can be lossy, and the rounding direction is contained in the result of these two methods. To use the default IEEE 754 rounding mode (rounding to nearest), the Repr::to_f32 and Repr::to_f64 methods can be used for convenience.

Convert from/to UBig/IBig

Converting from UBig and IBig is trivial and lossless through From. However, the reverse direction can be lossy.

The TryFrom trait and to_int() method are the two supported ways to convert from FBig to IBig. To convert to UBig, please first convert to IBig. When converting to IBig, TryFrom returns Ok only when the floating point number is not infinite and doesnt have fractional part. To convert with rounding, use to_int() instead.

Implementations

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn to_decimal(&self) -> Approximation<FBig<R, 10>, Rounding>

Convert the float number to base 10 (with decimal exponents).

Its equivalent to self.with_base::<10>(). See with_base() for the precision and rounding behavior.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::HalfAway, Rounding::*};

assert_eq!(
    FBig::<HalfAway, 2>::from_str_native("0x1234")?.to_decimal(),
    Exact(DBig::from_str_native("4660")?)
);
assert_eq!(
    FBig::<HalfAway, 2>::from_str_native("0x12.34")?.to_decimal(),
    Inexact(DBig::from_str_native("18.20")?, NoOp)
);
assert_eq!(
    FBig::<HalfAway, 2>::from_str_native("0x1.234p-4")?.to_decimal(),
    Inexact(DBig::from_str_native("0.07111")?, AddOne)
);

Panics

Panics if the associated context has unlimited precision and the conversion cannot be performed losslessly.

pub fn to_binary(&self) -> Approximation<FBig<R, 2>, Rounding>

Convert the float number to base 2 (with binary exponents).

Its equivalent to self.with_base::<2>(). See with_base() for the precision and rounding behavior.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::HalfAway, Rounding::*};

assert_eq!(
    DBig::from_str_native("1234")?.to_binary(),
    Exact(FBig::<HalfAway, 2>::from_str_native("0x4d2")?)
);
assert_eq!(
    DBig::from_str_native("12.34")?.to_binary(),
    Inexact(FBig::<HalfAway, 2>::from_str_native("0xc.57")?, NoOp)
);
assert_eq!(
    DBig::from_str_native("1.234e-1")?.to_binary(),
    Inexact(FBig::<HalfAway, 2>::from_str_native("0x1.f97p-4")?, NoOp)
);

Panics

Panics if the associated context has unlimited precision and the conversion cannot be performed losslessly.

pub fn with_precision( self, precision: usize ) -> Approximation<FBig<R, B>, Rounding>

Explicitly change the precision of the float number.

If the given precision is less than the current value in the context, it will be rounded with the rounding mode specified by the generic parameter.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::HalfAway, Rounding::*};

let a = DBig::from_str_native("2.345")?;
assert_eq!(a.precision(), 4);
assert_eq!(
    a.clone().with_precision(3),
    Inexact(DBig::from_str_native("2.35")?, AddOne)
);
assert_eq!(
    a.clone().with_precision(5),
    Exact(DBig::from_str_native("2.345")?)
);

pub fn with_rounding<NewR>(self) -> FBig<NewR, B>where NewR: Round,

Explicitly change the rounding mode of the number.

This operation doesnt modify the underlying representation, it only changes the rounding mode in the context.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::{HalfAway, Zero}, Rounding::*};

type DBigHalfAway = DBig;
type DBigZero = FBig::<Zero, 10>;

let a = DBigHalfAway::from_str_native("2.345")?;
let b = DBigZero::from_str_native("2.345")?;
assert_eq!(a.with_rounding::<Zero>(), b);

pub fn with_base<const NewB: u64>( self ) -> Approximation<FBig<R, NewB>, Rounding>

Explicitly change the base of the float number.

This function internally calls with_base_and_precision. The precision of the result number will be calculated in such a way that the new limit of the significand is less than or equal to before. That is, the new precision will be the max integer such that

NewB ^ new_precision <= B ^ old_precision

If any rounding happens during the conversion, it follows the rounding mode specified by the generic parameter.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::Zero, Rounding::*};

type FBin = FBig;
type FDec = FBig<Zero, 10>;
type FHex = FBig<Zero, 16>;

let a = FBin::from_str_native("0x1.234")?; // 0x1234 * 2^-12
assert_eq!(
    a.clone().with_base::<10>(),
    // 1.1376953125 rounded towards zero
    Inexact(FDec::from_str_native("1.137")?, NoOp)
);
assert_eq!(
    a.clone().with_base::<16>(),
    // conversion is exact when the new base is a power of the old base
    Exact(FHex::from_str_native("1.234")?)
);

Panics

Panics if the associated context has unlimited precision and the conversion cannot be performed losslessly.

pub fn with_base_and_precision<const NewB: u64>( self, precision: usize ) -> Approximation<FBig<R, NewB>, Rounding>

Explicitly change the base of the float number with given precision (under the new base).

Infinities are mapped to infinities inexactly, the error will be NoOp.

Conversion for float numbers with unlimited precision is only allowed in following cases:

The number is infinite

The new base NewB is a power of B

B is a power of the new base NewB

Examples

use dashu_base::Approximation::*;
use dashu_float::round::{mode::Zero, Rounding::*};

type FBin = FBig;
type FDec = FBig<Zero, 10>;
type FHex = FBig<Zero, 16>;

let a = FBin::from_str_native("0x1.234")?; // 0x1234 * 2^-12
assert_eq!(
    a.clone().with_base_and_precision::<10>(8),
    // 1.1376953125 rounded towards zero
    Inexact(FDec::from_str_native("1.1376953")?, NoOp)
);
assert_eq!(
    a.clone().with_base_and_precision::<16>(8),
    // conversion can be exact when the new base is a power of the old base
    Exact(FHex::from_str_native("1.234")?)
);
assert_eq!(
    a.clone().with_base_and_precision::<16>(2),
    // but the conversion is still inexact if the target precision is smaller
    Inexact(FHex::from_str_native("1.2")?, NoOp)
);

Panics

Panics if the associated context has unlimited precision and the conversion cannot be performed losslessly.

pub fn to_int(&self) -> Approximation<IBig, Rounding>

Convert the float number to integer with the given rounding mode.

Warning

If the float number has a very large exponent, it will be evaluated and result in allocating an huge integer and it might eat up all your memory.

To get a rough idea of how big the number is, its recommended to use EstimatedLog2.

Examples

use dashu_base::Approximation::*;
use dashu_float::round::Rounding::*;

assert_eq!(
    DBig::from_str_native("1234")?.to_int(),
    Exact(1234.into())
);
assert_eq!(
    DBig::from_str_native("1.234e6")?.to_int(),
    Exact(1234000.into())
);
assert_eq!(
    DBig::from_str_native("1.234")?.to_int(),
    Inexact(1.into(), NoOp)
);

Panics

Panics if the number is infinte

pub fn to_f32(&self) -> Approximation<f32, Rounding>

Convert the float number to f32 with HalfEven rounding mode regardless of the mode associated with this number.

Examples

assert_eq!(DBig::from_str_native("1.234")?.to_f32().value(), 1.234);
assert_eq!(DBig::INFINITY.to_f32().value(), f32::INFINITY);

pub fn to_f64(&self) -> Approximation<f64, Rounding>

Convert the float number to f64 with HalfEven rounding mode regardless of the mode associated with this number.

Examples

assert_eq!(DBig::from_str_native("1.234")?.to_f64().value(), 1.234);
assert_eq!(DBig::INFINITY.to_f64().value(), f64::INFINITY);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn powi(&self, exp: IBig) -> FBig<R, B>

Raise the floating point number to an integer power.

Examples

let a = DBig::from_str_native("-1.234")?;
assert_eq!(a.powi(10.into()), DBig::from_str_native("8.188")?);

pub fn powf(&self, exp: &FBig<R, B>) -> FBig<R, B>

Raise the floating point number to an floating point power.

Examples

let x = DBig::from_str_native("1.23")?;
let y = DBig::from_str_native("-4.56")?;
assert_eq!(x.powf(&y), DBig::from_str_native("0.389")?);

pub fn exp(&self) -> FBig<R, B>

Calculate the exponential function (e) on the floating point number.

Examples

let a = DBig::from_str_native("-1.234")?;
assert_eq!(a.exp(), DBig::from_str_native("0.2911")?);

pub fn exp_m1(&self) -> FBig<R, B>

Calculate the exponential minus one function (e-1) on the floating point number.

Examples

let a = DBig::from_str_native("-0.1234")?;
assert_eq!(a.exp_m1(), DBig::from_str_native("-0.11609")?);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn from_repr(repr: Repr<B>, context: Context<R>) -> FBig<R, B>

Create a FBig instance from Repr and Context.

This method should not be used in most cases. Its designed to be used when you hold a Repr instance and want to create an FBig from that.

Examples

use dashu_float::{Repr, Context};

assert_eq!(DBig::from_repr(Repr::one(), Context::new(1)), DBig::ONE);
assert_eq!(DBig::from_repr(Repr::infinity(), Context::new(1)), DBig::INFINITY);

Panics

Panics if the Repr has more digits than the precision limit specified in the context. Note that this condition is not checked in release build.

pub const ZERO: FBig<R, B> = Self::new(Repr::zero(), Context::new(0))

FBig with value 0 and unlimited precision

To test if the float number is zero, use self.repr().is_zero().

pub const ONE: FBig<R, B> = Self::new(Repr::one(), Context::new(0))

FBig with value 1 and unlimited precision

To test if the float number is one, use self.repr().is_one().

pub const NEG_ONE: FBig<R, B> = Self::new(Repr::neg_one(), Context::new(0))

FBig with value -1 and unlimited precision

pub const INFINITY: FBig<R, B> = Self::new(Repr::infinity(), Context::new(0))

FBig instance representing the positive infinity (+)

To test if the float number is infinite, use self.repr().infinite().

pub const NEG_INFINITY: FBig<R, B> = Self::new(Repr::neg_infinity(), Context::new(0))

FBig instance representing the negative infinity (-)

To test if the float number is infinite, use self.repr().infinite().

pub const fn precision(&self) -> usize

Get the maximum precision set for the float number.

Its equivalent to self.context().precision().

Examples

use dashu_float::Repr;

let a = DBig::from_str_native("1.234")?;
assert!(a.repr().significand() <= &Repr::<10>::BASE.pow(a.precision()));

pub fn digits(&self) -> usize

Get the number of the significant digits in the float number

Its equivalent to self.repr().digits().

This value is also the actual precision needed for the float number. Shrink to this value using with_precision() will not cause loss of float precision.

Examples

use dashu_base::Approximation::*;

let a = DBig::from_str_native("-1.234e-3")?;
assert_eq!(a.digits(), 4);
assert!(matches!(a.clone().with_precision(4), Exact(_)));
assert!(matches!(a.clone().with_precision(3), Inexact(_, _)));

pub const fn context(&self) -> Context<R>

Get the context associated with the float number

pub const fn repr(&self) -> &Repr<B>

Get a reference to the underlying numeric representation

pub fn into_repr(self) -> Repr<B>

Get the underlying numeric representation

Examples

use dashu_float::Repr;

let a = DBig::ONE;
assert_eq!(a.into_repr(), Repr::<10>::one());

pub fn from_parts(significand: IBig, exponent: isize) -> FBig<R, B>

Convert raw parts (significand, exponent) into a float number.

The precision will be inferred from significand (the lowest k such that significand <= base^k)

Examples

let a = DBig::from_parts((-1234).into(), -2);
assert_eq!(a, DBig::from_str_native("-12.34")?);
assert_eq!(a.precision(), 4); // 1234 has 4 (decimal) digits

pub const fn from_parts_const( sign: Sign, significand: u128, exponent: isize, min_precision: Option<usize> ) -> FBig<R, B>

Convert raw parts (significand, exponent) into a float number in a const context.

It requires that the significand fits in a DoubleWord.

The precision will be inferred from significand (the lowest k such that significand <= base^k). If the min_precision is provided, then the higher one from the given and inferred precision will be used as the final precision.

Examples

use dashu_base::Sign;

const A: DBig = DBig::from_parts_const(Sign::Negative, 1234, -2, None);
assert_eq!(A, DBig::from_str_native("-12.34")?);
assert_eq!(A.precision(), 4); // 1234 has 4 (decimal) digits

const B: DBig = DBig::from_parts_const(Sign::Negative, 1234, -2, Some(5));
assert_eq!(B.precision(), 5); // overrided by the argument

pub fn ulp(&self) -> FBig<R, B>

Return the value of the least significant digit of the float number x, such that x + ulp is the first float number greater than x (given the precision from the context).

Examples

assert_eq!(DBig::from_str_native("1.23")?.ulp(), DBig::from_str_native("0.01")?);
assert_eq!(DBig::from_str_native("01.23")?.ulp(), DBig::from_str_native("0.001")?);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn ln(&self) -> FBig<R, B>

Calculate the natural logarithm function (log(x)) on the float number.

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.ln(), DBig::from_str_native("0.2103")?);

pub fn ln_1p(&self) -> FBig<R, B>

Calculate the natural logarithm function (log(x+1)) on the float number

Examples

let a = DBig::from_str_native("0.1234")?;
assert_eq!(a.ln_1p(), DBig::from_str_native("0.11636")?);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn square(&self) -> FBig<R, B>

Compute the square of this number (self * self)

Examples

let a = DBig::from_str_native("-1.234")?;
assert_eq!(a.square(), DBig::from_str_native("1.523")?);

pub fn cubic(&self) -> FBig<R, B>

Compute the square of this number (self * self)

Examples

let a = DBig::from_str_native("-1.234")?;
assert_eq!(a.cubic(), DBig::from_str_native("-1.879")?);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn from_str_native(src: &str) -> Result<FBig<R, B>, ParseError>

Convert a string in the native base (i.e. radix B) to FBig.

If the parsing succeeded, the result number will be losslessly parsed from the input string.

This function is the actual implementation of the FromStr trait.

Note: Infinites are intentionally not supported by this function.

Format

The valid representations include

aaa or aaa.

aaa is represented in native base B without base prefixes.

aaa.bbb = aaabbb / base ^ len(bbb)

aaa and bbb are represented in native base B without base prefixes.

len(bbb) represents the number of digits in bbb, e.g len(bbb) is 3. (Same below)

aaa.bbb@cc = aaabbb * base ^ (cc - len(bbb))

aaa and bbb are represented in native base B

This is consistent with the representation used by GNU GMP.

aaa.bbbEcc = aaabbb * 10 ^ (cc - len(bbb))

E could be lower case, base B must be 10

aaa and bbb are all represented in decimal

0xaaa or 0xaaa

0xaaa.bbb = 0xaaabbb / 16 ^ len(bbb)

0xaaa.bbbPcc = 0xaaabbb / 16 ^ len(bbb) * 2 ^ cc

P could be lower case, base B must be 2 (not 16!)

aaa and bbb are represented in hexadecimal

This is consistent with the C++ hexadecimal literals.

aaa.bbbBcc = aaabbb * 2 ^ (cc - len(bbb))

aaa.bbbOcc = aaabbb * 8 ^ (cc - len(bbb))

aaa.bbbHcc = aaabbb * 16 ^ (cc - len(bbb))

B/O/H could be lower case, and base B must be consistent with the marker.

aaa and bbb are represented in binary/octal/hexadecimal correspondingly without prefix.

This is consistent with some scientific notations described in Wikipedia.

Digits 10-35 are represented by a-z or A-Z.

Literal aaa and cc above can be signed, but bbb must be unsigned. All cc are represented in decimal. Either aaa or bbb can be omitted when its value is zero, but they are not allowed to be omitted at the same time.

Precision

The precision of the parsed number is determined by the number of digits that are presented in the input string. For example, the numbers parsed from 12.34 or 1.234e-1 will have a precision of 4, while the ones parsed from 12.34000 or 00012.34 will have a precision of 7.

Examples

use dashu_base::Approximation::*;

let a = DBig::from_str_native("-1.23400e-3")?;
let b = DBig::from_str_native("-123.4@-05")?;
assert_eq!(a, b);
assert_eq!(a.precision(), 6);
assert_eq!(b.precision(), 4);

assert!(DBig::from_str_native("-0x1.234p-3").is_err());
assert!(DBig::from_str_native("-1.234H-3").is_err());

Panics

Panics if the base B is not between MIN_RADIX and MAX_RADIX inclusive.

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn sqrt(&self) -> FBig<R, B>

Calculate the square root of the floating point number.

Examples

let a = DBig::from_str_native("1.23")?;
assert_eq!(a.sqrt(), DBig::from_str_native("1.11")?);

impl<R, const B: u64> FBig<R, B>where R: Round,

pub fn trunc(&self) -> FBig<R, B>

Get the integral part of the float

See FBig::round for how the output precision is determined.

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.trunc(), DBig::from_str_native("1")?);
// the actual precision of the integral part is 1 digit
assert_eq!(a.trunc().precision(), 1);

Panics

Panics if the number is infinte

pub fn split_at_point(self) -> (FBig<R, B>, FBig<R, B>)

Split the rational number into integral and fractional parts (split at the radix point)

Its equivalent to (self.trunc(), self.fract())

Examples

let a = DBig::from_str_native("1.234")?;
let (trunc, fract) = a.split_at_point();
assert_eq!(trunc, DBig::from_str_native("1.0")?);
assert_eq!(fract, DBig::from_str_native("0.234")?);
// the actual precision of the fractional part is 3 digits
assert_eq!(trunc.precision(), 1);
assert_eq!(fract.precision(), 3);

pub fn fract(&self) -> FBig<R, B>

Get the fractional part of the float

Note: this function will adjust the precision accordingly!

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.fract(), DBig::from_str_native("0.234")?);
// the actual precision of the fractional part is 3 digits
assert_eq!(a.fract().precision(), 3);

Panics

Panics if the number is infinte

pub fn ceil(&self) -> FBig<R, B>

Returns the smallest integer greater than or equal to self.

See FBig::round for how the output precision is determined.

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.ceil(), DBig::from_str_native("2")?);

// works for very large exponent
let b = DBig::from_str_native("1.234e10000")?;
assert_eq!(b.ceil(), b);

Panics

Panics if the number is infinte

pub fn floor(&self) -> FBig<R, B>

Returns the largest integer less than or equal to self.

See FBig::round for how the output precision is determined.

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.floor(), DBig::from_str_native("1")?);

// works for very large exponent
let b = DBig::from_str_native("1.234e10000")?;
assert_eq!(b.floor(), b);

Panics

Panics if the number is infinte

pub fn round(&self) -> FBig<R, B>

Returns the integer nearest to self.

If there are two integers equally close, then the one farther from zero is chosen.

Examples

let a = DBig::from_str_native("1.234")?;
assert_eq!(a.round(), DBig::from_str_native("1")?);

// works for very large exponent
let b = DBig::from_str_native("1.234e10000")?;
assert_eq!(b.round(), b);

Precision

If self is an integer, the result will have the same precision as self. If self has fractional part, then the precision will be subtracted by the digits in the fractional part. Examples:

1.00e100 (precision = 3) rounds to 1.00e100 (precision = 3)

1.234 (precision = 4) rounds to 1. (precision = 1)

1.234e-10 (precision = 4) rounds to 0. (precision = 0, i.e arbitrary precision)

Panics

Panics if the number is infinte

impl<R, const B: u64> FBig<R, B>where R: Round,

pub const fn sign(&self) -> Sign

Get the sign of the number. Zero value has a positive sign.

Examples

assert_eq!(DBig::ZERO.sign(), Sign::Positive);
assert_eq!(DBig::from_str_native("-1.234")?.sign(), Sign::Negative);

pub const fn signum(&self) -> FBig<R, B>

A number representing the sign of self.

FBig::ONE if the number is positive (including inf)

FBig::ZERO if the number is zero

FBig::NEG_ONE if the number is negative (including -inf)

Examples

assert_eq!(DBig::from_str_native("2.01")?.signum(), DBig::ONE);
assert_eq!(DBig::from_str_native("-1.234")?.signum(), DBig::NEG_ONE);