Struct dashu::float::Repr

pub struct Repr<const BASE: u64> { /* private fields */ }

Underlying representation of an arbitrary precision floating number.

The floating point number is represented as significand * base^exponent, where the type of the significand is IBig, and the type of exponent is isize. The representation is always normalized (nonzero signficand is not divisible by the base, or zero signficand with zero exponent).

When it’s used together with a Context, its precision will be limited so that |signficand| < base^precision. However, the precision limit is not always enforced. In rare cases, the significand can have one more digit than the precision limit.

Infinity

This struct supports representing the infinity, but the infinity is only supposed to be used as sentinels. That is, only equality test and comparison are implemented for the infinity. Any other operations on the infinity will lead to panic. If an operation result is too large or too small, the operation will panic instead of returning an infinity.

Implementations

impl<const B: u64> Repr<B>

pub fn to_f32(&self) -> Approximation<f32, Rounding>

Convert the float number representation to a f32 with the default IEEE 754 rounding mode.

The default IEEE 754 rounding mode is HalfEven (rounding to nearest, ties to even). To convert the float number with a specific rounding mode, please use FBig::to_f32.

Examples

assert_eq!(Repr::<2>::one().to_f32(), Exact(1.0));
assert_eq!(Repr::<10>::infinity().to_f32(), Inexact(f32::INFINITY, NoOp));

pub fn to_f64(&self) -> Approximation<f64, Rounding>

Convert the float number representation to a f64 with the default IEEE 754 rounding mode.

The default IEEE 754 rounding mode is HalfEven (rounding to nearest, ties to even). To convert the float number with a specific rounding mode, please use FBig::to_f64.

Examples

assert_eq!(Repr::<2>::one().to_f64(), Exact(1.0));
assert_eq!(Repr::<10>::infinity().to_f64(), Inexact(f64::INFINITY, NoOp));

pub fn to_int(&self) -> Approximation<IBig, Rounding>

Convert the float number representation to a IBig.

The fractional part is always rounded to zero. To convert with other rounding modes, please use FBig::to_int().

Warning

If the float number has a very large exponent, it will be evaluated and result in allocating an huge integer and it might eat up all your memory.

To get a rough idea of how big the number is, it’s recommended to use EstimatedLog2.

Examples

assert_eq!(Repr::<2>::neg_one().to_int(), Exact(IBig::NEG_ONE));

Panics

Panics if the number is infinte.

impl<const B: u64> Repr<B>

pub fn from_str_native(src: &str) -> Result<(Repr<B>, usize), ParseError>

Convert a string in the native base (i.e. radix B) to Repr.

Upon success, this method returns an Repr and the number of digits (in radix B) contained in the string.

This method is the underlying implementation of FBig::from_str_native, see the docs for that function for details.

impl<const B: u64> Repr<B>

pub const BASE: IBig = base_as_ibig::<B>()

The base of the representation. It’s exposed as an IBig constant.

pub const fn zero() -> Repr<B>

Create a Repr instance representing value zero

pub const fn one() -> Repr<B>

Create a Repr instance representing value one

pub const fn neg_one() -> Repr<B>

Create a Repr instance representing value negative one

pub const fn infinity() -> Repr<B>

Create a Repr instance representing the (positive) infinity

pub const fn neg_infinity() -> Repr<B>

Create a Repr instance representing the negative infinity

pub const fn is_zero(&self) -> bool

Determine if the Repr represents zero

Examples

assert!(Repr::<2>::zero().is_zero());
assert!(!Repr::<10>::one().is_zero());

pub const fn is_one(&self) -> bool

Determine if the Repr represents one

Examples

assert!(Repr::<2>::zero().is_zero());
assert!(!Repr::<10>::one().is_zero());

pub const fn is_infinite(&self) -> bool

Determine if the Repr represents the (±)infinity

Examples

assert!(Repr::<2>::infinity().is_infinite());
assert!(Repr::<10>::neg_infinity().is_infinite());
assert!(!Repr::<10>::one().is_infinite());

pub const fn is_finite(&self) -> bool

Determine if the Repr represents a finite number

Examples

assert!(Repr::<2>::zero().is_finite());
assert!(Repr::<10>::one().is_finite());
assert!(!Repr::<16>::infinity().is_finite());

pub const fn sign(&self) -> Sign

Get the sign of the number

Examples

assert_eq!(Repr::<2>::zero().sign(), Sign::Positive);
assert_eq!(Repr::<2>::neg_one().sign(), Sign::Negative);
assert_eq!(Repr::<10>::neg_infinity().sign(), Sign::Negative);

pub fn digits(&self) -> usize

Get the number of digits (under base B) in the significand.

If the number is 0, then 0 is returned (instead of 1).

Examples

assert_eq!(Repr::<2>::zero().digits(), 0);
assert_eq!(Repr::<2>::one().digits(), 1);
assert_eq!(Repr::<10>::one().digits(), 1);

assert_eq!(Repr::<10>::new(100.into(), 0).digits(), 1); // 1e2
assert_eq!(Repr::<10>::new(101.into(), 0).digits(), 3);

pub fn digits_ub(&self) -> usize

Fast over-estimation of digits

Examples

assert_eq!(Repr::<2>::zero().digits_ub(), 0);
assert_eq!(Repr::<2>::one().digits_ub(), 1);
assert_eq!(Repr::<10>::one().digits_ub(), 1);
assert_eq!(Repr::<2>::new(31.into(), 0).digits_ub(), 5);
assert_eq!(Repr::<10>::new(99.into(), 0).digits_ub(), 2);

pub fn digits_lb(&self) -> usize

Fast under-estimation of digits

Examples

assert_eq!(Repr::<2>::zero().digits_lb(), 0);
assert_eq!(Repr::<2>::one().digits_lb(), 0);
assert_eq!(Repr::<10>::one().digits_lb(), 0);
assert!(Repr::<10>::new(1001.into(), 0).digits_lb() <= 3);

pub fn new(significand: IBig, exponent: isize) -> Repr<B>

Create a Repr from the significand and exponent. This constructor will normalize the representation.

Examples

let a = Repr::<2>::new(400.into(), -2);
assert_eq!(a.significand(), &IBig::from(25));
assert_eq!(a.exponent(), 2);

let b = Repr::<10>::new(400.into(), -2);
assert_eq!(b.significand(), &IBig::from(4));
assert_eq!(b.exponent(), 0);

pub fn significand(&self) -> &IBig

Get the significand of the representation

pub fn exponent(&self) -> isize

Get the exponent of the representation

pub fn into_parts(self) -> (IBig, isize)

Convert the float number into raw (signficand, exponent) parts

Examples

use dashu_int::IBig;

let a = Repr::<2>::new(400.into(), -2);
assert_eq!(a.into_parts(), (IBig::from(25), 2));

let b = Repr::<10>::new(400.into(), -2);
assert_eq!(b.into_parts(), (IBig::from(4), 0));