Struct dashu::integer::modular::Reduced

pub struct Reduced<'a>(/* private fields */);

Modular arithmetic.

Examples

let ring = ConstDivisor::new(UBig::from(10000u32));
let x = ring.reduce(12345);
let y = ring.reduce(55443);
assert_eq!((x - y).residue(), UBig::from(6902u32));

Implementations

impl<'a> Reduced<'a>

pub fn dbl(self) -> Reduced<'a>

Calculate 2*target mod m in reduced form

Examples

let p = UBig::from(0x1234u16);
let ring = ConstDivisor::new(p.clone());
let a = ring.reduce(4000);
assert_eq!(a.dbl(), ring.reduce(4000 + 4000));

impl Reduced<'_>

pub fn residue(&self) -> UBig

Get the residue in range 0..n in an n-element ring.

Examples

let ring = ConstDivisor::new(UBig::from(100u8));
let x = ring.reduce(-1234);
assert_eq!(x.residue(), UBig::from(66u8));

pub fn modulus(&self) -> UBig

Get the modulus of the ring that this element belongs to.

impl<'a> Reduced<'a>

pub fn inv(&self) -> Option<Reduced<'a>>

Multiplicative inverse.

Examples

// A Mersenne prime.
let p = UBig::from(2u8).pow(127) - UBig::ONE;
let ring = ConstDivisor::new(p.clone());
// Fermat's little theorem: a^(p-2) = a^-1 (mod p)
let a = ring.reduce(123);
let ainv = a.clone().inv().unwrap();
assert_eq!(ainv, a.pow(&(p - UBig::from(2u8))));
assert_eq!((a * ainv).residue(), UBig::ONE);

impl<'a> Reduced<'a>

pub fn sqr(&self) -> Reduced<'a>

Calculate target^2 mod m in reduced form

Examples

let p = UBig::from(0x1234u16);
let ring = ConstDivisor::new(p.clone());
let a = ring.reduce(4000);
assert_eq!(a.sqr(), ring.reduce(4000 * 4000));

impl<'a> Reduced<'a>

pub fn pow(&self, exp: &UBig) -> Reduced<'a>

Exponentiation.

If you want use a negative exponent, you can first use inv() to convert the base to its inverse, and then call this method.

Examples

// A Mersenne prime.
let p = UBig::from(2u8).pow(607) - UBig::ONE;
let ring = ConstDivisor::new(p.clone());
// Fermat's little theorem: a^(p-1) = 1 (mod p)
let a = ring.reduce(123);
assert_eq!(a.pow(&(p - UBig::ONE)), ring.reduce(1));