Struct dashu::integer::IBig

#[repr(transparent)]
pub struct IBig(_);

An signed arbitrary precision integer.

This struct represents an arbitrarily large signed integer. Technically the size of the integer is bounded by the memory size, but its enough for practical use on modern devices.

Parsing and printing

There are four ways to create an IBig instance:

Use predifined constants (e.g. IBig::ZERO, IBig::NEG_ONE).

Use the literal macro ibig! defined in the dashu-macro crate.

Construct from a Sign and a UBig instance.

Parse from a string.

Parsing from either literal or string supports representation with base 2~36.

For printing, the IBig type supports common formatting traits (Display, Debug, LowerHex, etc.). Specially, printing huge number using Debug will conveniently omit the middle digits of the number, only print the least and most significant (decimal) digits.

// parsing
let a = IBig::from(408580953453092208335085386466371u128);
let b = IBig::from(-0x1231abcd4134i64);
let c = IBig::from_str_radix("a2a123bbb127779cccc123", 32)?;
let d = IBig::from_str_radix("-1231abcd4134", 16)?;
assert_eq!(a, c);
assert_eq!(b, d);

// printing
assert_eq!(format!("{}", IBig::from(12)), "12");
assert_eq!(format!("{:#X}", IBig::from(-0xabcd)), "-0xABCD");
if Word::BITS == 64 {
    // number of digits to display depends on the word size
    assert_eq!(
        format!("{:?}", IBig::NEG_ONE << 1000),
        "-1071508607186267320..4386837205668069376"
    );
}

Memory

The internal representation of IBig is exactly the same as UBig. It just use a small trick to store the sign bit without additional memory allocation. This means that IBig also has the small integer optimization and the niche bit to use with simple enums.

use core::mem::size_of;
assert_eq!(size_of::<IBig>(), size_of::<UBig>());
assert_eq!(size_of::<IBig>(), size_of::<Option<IBig>>());

Implementations

impl IBig

pub fn trailing_zeros(&self) -> Option<usize>

Returns the number of trailing zeros in the twos complement binary representation.

In other words, it is the largest n such that 2 to the power of n divides the number.

For 0, it returns None.

Examples

assert_eq!(IBig::from(17).trailing_zeros(), Some(0));
assert_eq!(IBig::from(-48).trailing_zeros(), Some(4));
assert_eq!(IBig::from(-0b101000000).trailing_zeros(), Some(6));
assert_eq!(IBig::ZERO.trailing_zeros(), None);

pub fn trailing_ones(&self) -> Option<usize>

Returns the number of trailing ones in the twos complement binary representation.

For positive self, its equivalent to self.unsigned_abs().trailing_zeros(). For negative self, its equivalent to (!self.unsigned_abs() + 1).trailing_zeros().

For -1, it returns None.

Examples

assert_eq!(IBig::from(17).trailing_ones(), Some(1));
assert_eq!(IBig::from(-48).trailing_ones(), Some(0));
assert_eq!(IBig::from(-0b101000001).trailing_ones(), Some(6));
assert_eq!(IBig::NEG_ONE.trailing_ones(), None);

impl IBig

pub fn to_f32(&self) -> Approximation<f32, Sign>

Convert to f32.

Round to nearest, breaking ties to even last bit. The returned approximation is exact if the integer is exactly representable by f32, otherwise the error field of the approximation contains the sign of result - self.

Examples

assert_eq!(IBig::from(-134).to_f32().value(), -134.0f32);

pub fn to_f64(&self) -> Approximation<f64, Sign>

Convert to f64.

Round to nearest, breaking ties to even last bit. The returned approximation is exact if the integer is exactly representable by f64, otherwise the error field of the approximation contains the sign of result - self.

Examples

assert_eq!(IBig::from(-134).to_f64().value(), -134.0f64);

impl IBig

pub fn in_radix(&self, radix: u32) -> InRadix<'_>

Representation in a given radix.

Panics

Panics if radix is not between 2 and 36 inclusive.

Examples

assert_eq!(format!("{}", IBig::from(-83).in_radix(3)), "-10002");
assert_eq!(format!("{:010}", IBig::from(-35).in_radix(36)), "-00000000z");

impl IBig

pub fn as_sign_words(&self) -> (Sign, &[u64])

Get the raw representation in Words.

If the number is zero, then empty slice will be returned.

Examples

assert_eq!(IBig::ZERO.as_sign_words(), (Sign::Positive, &[] as &[_]));
assert_eq!(IBig::NEG_ONE.as_sign_words().0, Sign::Negative);
assert_eq!(IBig::NEG_ONE.as_sign_words().1, &[1]);

pub const fn sign(&self) -> Sign

Get the sign of the number. Zero value has a positive sign.

Examples

assert_eq!(IBig::ZERO.sign(), Sign::Positive);
assert_eq!(IBig::from(2).sign(), Sign::Positive);
assert_eq!(IBig::from(-3).sign(), Sign::Negative);

pub fn into_parts(self) -> (Sign, UBig)

Convert the IBig into its Sign and UBig magnitude

Examples

assert_eq!(IBig::ZERO.into_parts(), (Sign::Positive, UBig::ZERO));
assert_eq!(IBig::ONE.into_parts(), (Sign::Positive, UBig::ONE));
assert_eq!(IBig::NEG_ONE.into_parts(), (Sign::Negative, UBig::ONE));

pub fn from_parts(sign: Sign, magnitude: UBig) -> IBig

Create an IBig from the Sign and UBig magnitude

Examples

assert_eq!(IBig::from_parts(Sign::Positive, UBig::ZERO), IBig::ZERO);
assert_eq!(IBig::from_parts(Sign::Positive, UBig::ONE), IBig::ONE);
assert_eq!(IBig::from_parts(Sign::Negative, UBig::ONE), IBig::NEG_ONE);

pub const fn from_parts_const(sign: Sign, dword: u128) -> IBig

Create an IBig in a const context.

The magnitude is limited to a DoubleWord.

Examples

const ONE: IBig = IBig::from_parts_const(Sign::Positive, 1);
assert_eq!(ONE, IBig::ONE);
const NEG_ONE: IBig = IBig::from_parts_const(Sign::Negative, 1);
assert_eq!(NEG_ONE, IBig::NEG_ONE);

pub const ZERO: IBig = Self(Repr::zero())

IBig with value 0

pub const ONE: IBig = Self(Repr::one())

IBig with value 1

pub const NEG_ONE: IBig = Self(Repr::neg_one())

IBig with value -1

pub const fn is_zero(&self) -> bool

Check whether the number is 0

Examples

assert!(IBig::ZERO.is_zero());
assert!(!IBig::ONE.is_zero());

pub const fn is_one(&self) -> bool

Check whether the number is 1

Examples

assert!(!IBig::ZERO.is_one());
assert!(IBig::ONE.is_one());

impl IBig

pub fn ilog(&self, base: &UBig) -> usize

Calculate the (truncated) logarithm of the magnitude of IBig

This function could takes a long time when the integer is very large. In applications where an exact result is not necessary, log2_bounds could be used.

Panics

Panics if the number is 0, or the base is 0 or 1

Examples

let base = UBig::from(3u8);
assert_eq!(IBig::from(-81).ilog(&base), 4);
assert_eq!(IBig::from(-1000).ilog(&base), 6);

impl IBig

pub fn square(&self) -> UBig

Compute the square of the number (self * self).

Examples

assert_eq!(IBig::from(-3).square(), UBig::from(9u8));

pub fn cubic(&self) -> IBig

Compute the cubic of the number (self * self * self).

Examples

assert_eq!(IBig::from(-3).cubic(), IBig::from(-27));

impl IBig

pub fn from_str_radix(src: &str, radix: u32) -> Result<IBig, ParseError>

Convert a string in a given base to IBig.

The string may contain a + or - prefix. Digits 10-35 are represented by a-z or A-Z.

Examples

assert_eq!(IBig::from_str_radix("-7ab", 32)?, IBig::from(-7499));

pub fn from_str_with_radix_prefix(src: &str) -> Result<(IBig, u32), ParseError>

Convert a string with an optional radix prefix to IBig, return the parsed integer and radix.

src may contain an + or - prefix before the radix prefix.

Allowed prefixes: 0b for binary, 0o for octal, 0x for hexadecimal.

Examples

assert_eq!(IBig::from_str_with_radix_prefix("+0o17")?, (IBig::from(0o17), 8));
assert_eq!(IBig::from_str_with_radix_prefix("-0x1f")?.0, IBig::from(-0x1f));

impl IBig

pub fn pow(&self, exp: usize) -> IBig

Raises self to the power of exp.

Examples

assert_eq!(IBig::from(-3).pow(3), IBig::from(-27));

impl IBig

pub fn nth_root(&self, n: usize) -> IBig

Calculate the nth-root of the integer rounding towards zero

Examples

assert_eq!(IBig::from(4).nth_root(2), IBig::from(2));
assert_eq!(IBig::from(-4).nth_root(3), IBig::from(-1));
assert_eq!(IBig::from(-1024).nth_root(5), IBig::from(-4));

Panics

If n is zero, or if n is even when the integer is negative.

impl IBig

pub const fn signum(&self) -> IBig

A number representing the sign of self.

IBig::ONE if the number is positive

IBig::ZERO if the number is zero

IBig::NEG_ONE if the number is negative

Examples

assert_eq!(IBig::from(-500).signum(), IBig::from(-1));