Struct dashu::integer::UBig

#[repr(transparent)]
pub struct UBig(_);

An unsigned arbitrary precision integer.

This struct represents an arbitrarily large unsigned integer. Technically the size of the integer is bounded by the memory size, but its enough for practical use on modern devices.

Parsing and printing

To create a UBig instance, there are three ways:

Use predifined constants (e.g. UBig::ZERO, UBig::ONE).

Use the literal macro ubig! defined in the dashu-macro crate.

Parse from a string.

Parsing from either literal or string supports representation with base 2~36.

For printing, the UBig type supports common formatting traits (Display, Debug, LowerHex, etc.). Specially, printing huge number using Debug will conveniently omit the middle digits of the number, only print the least and most significant (decimal) digits.

// parsing
let a = UBig::from(408580953453092208335085386466371u128);
let b = UBig::from(0x1231abcd4134u64);
let c = UBig::from_str_radix("a2a123bbb127779cccc123", 32)?;
let d = UBig::from_str_radix("1231abcd4134", 16)?;
assert_eq!(a, c);
assert_eq!(b, d);

// printing
assert_eq!(format!("{}", UBig::from(12u8)), "12");
assert_eq!(format!("{:#X}", UBig::from(0xabcdu16)), "0xABCD");
if Word::BITS == 64 {
    // number of digits to display depends on the word size
    assert_eq!(
        format!("{:?}", UBig::ONE << 1000),
        "1071508607186267320..4386837205668069376"
    );
}

Memory

Integers that fit in a DoubleWord will be inlined on stack and no heap allocation will be invoked. For large integers, they will be represented as an array of Words, and stored on heap.

Note that the UBig struct has a niche bit, therefore it can be used within simple enums with no memory overhead.

use core::mem::size_of;
assert_eq!(size_of::<UBig>(), size_of::<Option<UBig>>());

Implementations

impl UBig

pub fn set_bit(&mut self, n: usize)

Set the n-th bit, n starts from 0.

Examples

let mut a = UBig::from(0b100u8);
a.set_bit(0);
assert_eq!(a, UBig::from(0b101u8));
a.set_bit(10);
assert_eq!(a, UBig::from(0b10000000101u16));

pub fn clear_bit(&mut self, n: usize)

Clear the n-th bit, n starts from 0.

Examples

let mut a = UBig::from(0b101u8);
a.clear_bit(0);
assert_eq!(a, UBig::from(0b100u8));

pub fn trailing_zeros(&self) -> Option<usize>

Returns the number of trailing zeros in the binary representation.

In other words, it is the largest n such that 2 to the power of n divides the number.

For 0, it returns None.

Examples

assert_eq!(UBig::from(17u8).trailing_zeros(), Some(0));
assert_eq!(UBig::from(48u8).trailing_zeros(), Some(4));
assert_eq!(UBig::from(0b101000000u16).trailing_zeros(), Some(6));
assert_eq!(UBig::ZERO.trailing_zeros(), None);

pub fn trailing_ones(&self) -> Option<usize>

Returns the number of trailing ones in the binary representation.

In other words, it is the number of trailing zeros of it added by one.

This method never returns None.

Examples

assert_eq!(UBig::from(17u8).trailing_ones(), Some(1));
assert_eq!(UBig::from(48u8).trailing_ones(), Some(0));
assert_eq!(UBig::from(0b101001111u16).trailing_ones(), Some(4));
assert_eq!(UBig::ZERO.trailing_ones(), Some(0));

pub fn split_bits(self, n: usize) -> (UBig, UBig)

Split this integer into low bits and high bits.

Its returns are equal to (self & ((1 << n) - 1), self >> n).

Examples

let (lo, hi) = UBig::from(0b10100011u8).split_bits(4);
assert_eq!(hi, UBig::from(0b1010u8));
assert_eq!(lo, UBig::from(0b0011u8));

let x = UBig::from(0x90ffff3450897234u64);
let (lo, hi) = x.clone().split_bits(21);
assert_eq!(hi, (&x) >> 21);
assert_eq!(lo, x & ((UBig::ONE << 21) - 1u8));

pub fn clear_high_bits(&mut self, n: usize)

Clear the high bits from n+1-th bit.

This operation is equivalent to getting the lowest n bits on the integer i.e. self &= ((1 << n) - 1).

Examples

let mut x = UBig::from(0b10100011u8);
x.clear_high_bits(4);
assert_eq!(x, UBig::from(0b0011u8));

let mut x = UBig::from(0x90ffff3450897234u64);
let lo = (&x) & ((UBig::ONE << 21) - 1u8);
x.clear_high_bits(21);
assert_eq!(x, lo);

pub fn count_ones(&self) -> usize

Count the 1 bits in the integer

Examples

assert_eq!(UBig::from(0b10100011u8).count_ones(), 4);
assert_eq!(UBig::from(0x90ffff3450897234u64).count_ones(), 33);

let x = (UBig::ONE << 150) - 1u8;
assert_eq!(x.count_ones(), x.bit_len());

pub fn count_zeros(&self) -> Option<usize>

Count the 0 bits in the integer after the leading bit 1.

If the integer is zero, None will be returned.

Examples

assert_eq!(UBig::from(0b10100011u8).count_zeros(), Some(4));
assert_eq!(UBig::from(0x90ffff3450897234u64).count_zeros(), Some(31));

let x = (UBig::ONE << 150) - 1u8;
assert_eq!(x.count_zeros(), Some(0));

impl UBig

pub fn from_le_bytes(bytes: &[u8]) -> UBig

Construct from little-endian bytes.

Examples

assert_eq!(UBig::from_le_bytes(&[3, 2, 1]), UBig::from(0x010203u32));

pub fn from_be_bytes(bytes: &[u8]) -> UBig

Construct from big-endian bytes.

Examples

assert_eq!(UBig::from_be_bytes(&[1, 2, 3]), UBig::from(0x010203u32));

pub fn to_le_bytes(&self) -> Box<[u8], Global>

Return little-endian bytes.

Examples

assert!(UBig::ZERO.to_le_bytes().is_empty());
assert_eq!(*UBig::from(0x010203u32).to_le_bytes(), [3, 2, 1]);

pub fn to_be_bytes(&self) -> Box<[u8], Global>

Return big-endian bytes.

Examples

assert!(UBig::ZERO.to_be_bytes().is_empty());
assert_eq!(*UBig::from(0x010203u32).to_be_bytes(), [1, 2, 3]);

pub fn to_f32(&self) -> Approximation<f32, Sign>

Convert to f32.

Round to nearest, breaking ties to even last bit. The returned approximation is exact if the integer is exactly representable by f32, otherwise the error field of the approximation contains the sign of result - self.

Examples

assert_eq!(UBig::from(134u8).to_f32().value(), 134.0f32);

pub fn to_f64(&self) -> Approximation<f64, Sign>

Convert to f64.

Round to nearest, breaking ties to even last bit. The returned approximation is exact if the integer is exactly representable by f64, otherwise the error field of the approximation contains the sign of result - self.

Examples

assert_eq!(UBig::from(134u8).to_f64().value(), 134.0f64);

impl UBig

pub fn in_radix(&self, radix: u32) -> InRadix<'_>

Representation in a given radix.

Panics

Panics if radix is not between 2 and 36 inclusive.

Examples

assert_eq!(format!("{}", UBig::from(83u8).in_radix(3)), "10002");
assert_eq!(format!("{:+010}", UBig::from(35u8).in_radix(36)), "+00000000z");

impl UBig

pub fn ilog(&self, base: &UBig) -> usize

Calculate the (truncated) logarithm of the UBig

This function could takes a long time when the integer is very large. In applications where an exact result is not necessary, log2_bounds could be used.

Panics

Panics if the number is 0, or the base is 0 or 1

Examples

let base = UBig::from(3u8);
assert_eq!(UBig::from(81u8).ilog(&base), 4);
assert_eq!(UBig::from(1000u16).ilog(&base), 6);

impl UBig

pub fn square(&self) -> UBig

Calculate the square of the number (x * x).

Examples

assert_eq!(UBig::from(3u8).square(), UBig::from(9u8));

pub fn cubic(&self) -> UBig

Compute the cubic of the number (self * self * self).

Examples

assert_eq!(UBig::from(3u8).cubic(), UBig::from(27u8));

impl UBig

pub fn from_str_radix(src: &str, radix: u32) -> Result<UBig, ParseError>

Convert a string in a given base to UBig.

src may contain an optional + prefix. Digits 10-35 are represented by a-z or A-Z.

Examples

assert_eq!(UBig::from_str_radix("+7ab", 32)?, UBig::from(7499u16));

pub fn from_str_with_radix_prefix(src: &str) -> Result<(UBig, u32), ParseError>

Convert a string with an optional radix prefix to UBig, returns the parsed integer and radix.

src may contain an optional + before the radix prefix.

Allowed prefixes: 0b for binary, 0o for octal, 0x for hexadecimal.

Examples

assert_eq!(UBig::from_str_with_radix_prefix("+0o17")?, (UBig::from(0o17u8), 8));
assert_eq!(UBig::from_str_with_radix_prefix("0x1f")?.0, UBig::from(0x1fu8));

impl UBig

pub fn pow(&self, exp: usize) -> UBig

Raises self to the power of exp.

Examples

assert_eq!(UBig::from(3u8).pow(3), UBig::from(27u8));

impl UBig

pub fn remove(&mut self, factor: &UBig) -> Option<usize>

Divide out all multiples of the factor from the integer, returns the exponent of the removed factor.

For self = 0 or factor = 0 or 1, this method returns None.

impl UBig

pub fn nth_root(&self, n: usize) -> UBig

Calculate the nth-root of the integer rounding towards zero

Examples

assert_eq!(UBig::from(4u8).nth_root(2), UBig::from(2u8));
assert_eq!(UBig::from(4u8).nth_root(3), UBig::from(1u8));
assert_eq!(UBig::from(1024u16).nth_root(5), UBig::from(4u8));

Panics

If n is zero

impl UBig

pub const ZERO: UBig = Self(Repr::zero())

UBig with value 0

pub const ONE: UBig = Self(Repr::one())

UBig with value 1

pub fn as_words(&self) -> &[u64]

Get the raw representation in Words.

If the number is zero, then empty slice will be returned.

Examples

assert_eq!(UBig::ZERO.as_words(), &[] as &[Word]);
assert_eq!(UBig::ONE.as_words(), &[1]);

pub const fn from_word(word: u64) -> UBig

Create a UBig from a single Word.

Examples

const ZERO: UBig = UBig::from_word(0);
assert_eq!(ZERO, UBig::ZERO);
const ONE: UBig = UBig::from_word(1);
assert_eq!(ONE, UBig::ONE);

pub const fn from_dword(dword: u128) -> UBig

Create a UBig from a DoubleWord.

Examples

const ZERO: UBig = UBig::from_dword(0);
assert_eq!(ZERO, UBig::ZERO);
const ONE: UBig = UBig::from_dword(1);
assert_eq!(ONE, UBig::ONE);

pub fn from_words(words: &[u64]) -> UBig

Convert a sequence of Words into a UBig

Examples

assert_eq!(UBig::from_words(&[] as &[Word]), UBig::ZERO);
assert_eq!(UBig::from_words(&[1]), UBig::ONE);
assert_eq!(UBig::from_words(&[1, 1]), (UBig::ONE << Word::BITS as usize) + UBig::ONE);

pub const fn is_zero(&self) -> bool

Check whether the value is 0

Examples

assert!(UBig::ZERO.is_zero());
assert!(!UBig::ONE.is_zero());

pub const fn is_one(&self) -> bool

Check whether the value is 1

Examples

assert!(!UBig::ZERO.is_one());
assert!(UBig::ONE.is_one());