Module dashu::base

Defintions of common traits

Modules

approx

Trait definitions for approximated values

bit

Trait definitions for bitwise operations.

error

Error types.

math

Trait definitions for math operations

ring

Trait definitions for operations related to rings (integer/polynomial/etc.)

sign

Trait definitions for sign related operations.

Enums

Approximation

Represent an calculation result with a possible error.

ConversionError

Number out of bounds.

ParseError

Error parsing a number.

Sign

An enum representing the sign of a number

Traits

Abs

Absolute value.

AbsCmp

Compare the magnitude of this number to the magnitude of the other number

AbsEq

Check whether the magnitude of this number is equal the magnitude of the other number

BitTest

Bit query for integers

CubicRoot

Compute the cubic root of the number

CubicRootRem

Computer the floored cubic root of the number and return the remainder at the same time.

DivEuclid

Compute Euclidean quotient.

DivRem

Compute quotient and remainder at the same time.

DivRemAssign

Compute quotient inplace and return remainder at the same time.

DivRemEuclid

Compute Euclidean quotient and remainder at the same time.

EstimatedLog2

Fast estimation of the binary logarithm of a number

ExtendedGcd

Compute the greatest common divisor between self and the other operand, and return both the common divisor g and the Bézout coefficients respectively.

FloatEncoding

Support encoding and decoding of floats into (mantissa, exponent) parts.

Gcd

Compute the greatest common divisor.

Inverse

Compute the multiplicative inverse (aka. reciprocal) of the number.

PowerOfTwo

Functions related to the power of two.

RemEuclid

Compute Euclidean remainder.

Signed

This trait marks the number is signed.

SquareRoot

Compute the square root of the number

SquareRootRem

Computer the floored square root of the number and return the remainder at the same time.

UnsignedAbs

Unsigned absolute value.